+++ src/model/Product.java
@@ -5,7 +5,7 @@ public class Product {
     private int id;
     private String name;
     private String description;
-    private double price;
+    private int price; // đổi từ double sang int
     private int quantity;
     private String size;
     private String color;
@@ -14,7 +14,7 @@ public class Product {
     private String categoryName; // Optional: join từ bảng categories
     private String material;
 
-    public Product(int id, String name, String description, double price, int quantity,
+    public Product(int id, String name, String description, int price, int quantity, // đổi ở đây
             String size, String color, String material, String imageUrl, int categoryId, String categoryName) {
         this.id = id;
         this.name = name;
@@ -57,11 +57,11 @@ public class Product {
         this.description = description;
     }
 
-    public double getPrice() {
+    public int getPrice() { // đổi kiểu trả về
         return price;
     }
 
-    public void setPrice(double price) {
+    public void setPrice(int price) { // đổi kiểu tham số
         this.price = price;
     }
 
@@ -120,5 +120,4 @@ public class Product {
     public void setCategoryName(String categoryName) {
         this.categoryName = categoryName;
     }
-
 }
